+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added new menu items
 - added tbl_energy_storage_containers_firecontrols to myems_system_db in database
 - added tbl_energy_storage_containers_hvacs to myems_system_db in database
--
+- added tbl_spaces_energy_flow_diagrams to myems_system_db in database
 ### Changed
 - updated energy storage power station reportings
 - changed meter search box size to small